chore: extract PathNodes to it own file
